Make `ResponseParser` middleware fallback to body
When the `ResponseParser` middleware encounters invalid JSON or XML,
fall back to returning the original `body`.

In the case of `Nokogiri::XML`, invalid XML contents will be parsed into
an empty `Document` instance. Instead of relying on an exception to be
raised, the middleware validates that document and falls back when any
of its errors are syntax related.